feat(app): Move quotes to its own component
